Sprint 9 (#203)
* Sprint 9 (#201)

* Exam Type Admin Panels

Client testing found that examtype admin panels would be required if new exam types were to be added in the future. This page is only accessable by users with the SUPPORT role, and can create/delete exam types, as well as editing ones that currently exist.

* Pesticide Exams - Provide Functionality for Designate

Pesticide exams needed to be broken out of the "Add Non-ITA Exam" button such that a pesticide designate liaison could add these exams on their own. This button that is viewable on the exam inventory page can only be seen by csrs with the field pesticide_designate set to 1. This Add Exam functionality mirrors the functionality of the Non-ITA exam creation, but incorperates a select office chooser as well, and displays a notes field on the summary. This functionality has been tested with the GA role and LIAISON role, both with and without the pesticide_designate column set to 1 to ensure that this functionality is completely on its own.

* Office Drop Down Functionality

During testing of the re-enabled pesticide exam button, it was found that the opffice drop down selector component was only working for users with the liaison role code. This was functionality was modified to include users who have the pesticide_designate indicator field set to 1.

* Appointments Feature Flagging

During client testing, it was found out that the Appointments Feature needed to be separate from the Exams Feature flagging, due to some (not all) offices using the appointments feature. This PR adds a appointments_enabled_ind field to the office model/schema which controls whether or not the the feature shows up in theq nav menu or not.

* Bookings Contact Information

Clients have determined through testing that bookings for exams (all except group) and non-exam events, require contact information to be collected from exam writers. This PR adds the backend changes required to handle the collection of booking contact information to the bookings model/schema.

* Contact Info for Non Group Exam Bookings

During client testing, it was found that when non-group exam and non exam events were booked, that client contact information needed to be collected. This required a new field to be added to the booking model and front end changes to be made to the confirm booking and edit booking modals such that these details could be gathered from clients.

* Timezones

During client testing, it was found that due to group exams being created by GA's that might not belong to the office that they're creating the exam form (in the same timezone), that timezones needed to be dealt with during group exam creation. To accomplish this, a timezone model/schema was created and attached to the office model/schema through foreign key. This would allow users to choose one of the three timezones available in BC, and attach it to a specific office through the flask admin panels for theq application. This will allow front end development to easily create a group exam in the timezone of the office that the exam belongs to.

* SQL Script Clean-up

During testing of new back-end features, it was found out that the pre-population scripts for offices was incomplete. This PR addresses adding the 'appointments_enabled' field and initializing each office in the table with a value of 0.

* Implemented Time Tracking Feature

Applied router navigation guard to dash.vue so that when navigating away from theQ to either Exam Management or Room Booking, the AddCitizenModal appears (simplified version) pre-filtered to "Exams" catagory of service_reqs.  User can select a catagory and proceed, and can finish time tracking by either clicking the icon which appears or returning to the Q.

* Liaison Permission Strucutre - Back End

This PR addresses the back end changes required to impliment the new liaison permission that is required to remove the liaison role. This PR adds the migration file for the backend changes on the csr model/schema as well as the admin panel changes in order for users to add/edit this permission to other users.

* Liaison Permission Structure - End Points

This commit addresses the end-points for exams and bookings that rely on the LIAISON role. Now these end points look at the csr object created and whether or not they have the liaison_designate field set to 1 or not.

* Liaison Permission Structure - Front End

This PR addresses the front end changes required to allow liaison designates to view/create/edit/book group exams.

* ITA Permission Structure - Front End

Previously implemented ita_designate has been implemented on the front end now in this PR. According to the clients, the only things that being an ita_designate should affect are the ability to delete exams and the ability to see notifications.

* ITA Designate Getter Re-factor

During review/testing of previous PRs, it was found that the ita designate getter should return a boolean value, and be renamed to include an 'is'. The getter is now named 'is_ita_designate' and returns a true or false value. This refactor also includes modifying all files that used this getter previously, modifiying the name, and how it was used. This re-factor has been full tested to ensure that functionality has not been lost.

* Liaison Designate Re-factor

During review/testing of previous PRs, it was found that the liaison designate getter should return a boolean value, and be renamed to include an 'is'. The getter is now named 'is_liaison_designate' and returns a true or false value. This refactor also includes modifying all files that used this getter previously, modifiying the name, and how it was used. This re-factor has been full tested to ensure that functionality has not been lost.

* Financial Designate Re-factor

During review/testing of previous PRs, it was found that the financial designate getter should return a boolean value, and be renamed to include an 'is'. The getter is now named 'is_financial_designate' and returns a true or false value. This refactor also includes modifying all files that used this getter previously, modifiying the name, and how it was used. This re-factor has been full tested to ensure that functionality has not been lost.

* Pesticide Designate Re-factor

During review/testing of previous PRs, it was found that the pesticide designate getter should return a boolean value, and be renamed to include an 'is'. The getter is now named 'is_pesticide_designate' and returns a true or false value. This refactor also includes modifying all files that used this getter previously, modifiying the name, and how it was used. This re-factor has been full tested to ensure that functionality has not been lost.

* Fixed issue where navigation to Admin Panel was not possible. (#202)

Navigation guard on dash.vue was blocking navigation away.  Fixed issue.`