Booking Calendar - Booking Events in the Past
Client testing found that bookings could be done in the past, which should not be allowed. This PR corrects this by adding logic to the config object of the calendar, checking to see if the start date of the event being created is before a datetime object called "Today" which goes down the minute value of the datetime object created. Example: If the time of the object created is 3:03pm of that day, the calendar will not allow you to create an event until the 3:30pm-4:00pm block.