Sprint 10 (#215)
* Pesticide Exam Modal - Preset/Validate Exam for Pesticide Exam Type

Client testing found that the Pesticide Exam modal needed the exam type value to be preset to "Pesticide". This PR contains the fix for this issue by pre-setting the exam_type_id once the modal is opened, and validating the exam_type_id as well, such that the user can continue through the steps for creating an exam.

* Admin Panel - Designate Label Re-name

Client testing found that ITA, Pesticide, and Liaison designates needed to be re-named to properly reflect the csr's job title. This has been addressed in this PR.

- ITA -> Office Exam Manager
- Pesticide -> Pesticide Client Liaison/Program Specialist
- Liaison -> ITA Liaison/Program Specialist

* Responsive Exam Inventory Table

Client testing found that if a user with vision impairment zooms in on the exam inventory table, that it didn't have a horizontal scroll automatically enabled. This was fixed by adding the 'responsive' prop to the b-table on the exam-inventory table component, and tested by zooming on the component to see whether or not a scroll bar became active if the component was larger than the actual screensize of the user.

* Booking Calendar - Booking Events in the Past

Client testing found that bookings could be done in the past, which should not be allowed. This PR corrects this by adding logic to the config object of the calendar, checking to see if the start date of the event being created is before a datetime object called "Today" which goes down the minute value of the datetime object created. Example: If the time of the object created is 3:03pm of that day, the calendar will not allow you to create an event until the 3:30pm-4:00pm block.

* Group Exam Time Select

Client testing found the "Ok" confirm at the bottom of the Exam Time select drop down was not necessary. This PR seeing the confirm option on the Time Question component removed to remedy this.

* Input Field - Disable Autocomplete

Client testing found that vue input fields should have autocomplete disabled.

This was completed in a previous sprint, but developer testing found one field on the edit booking modal that didn't have this setting enabled.

* Generate CSV - GA Reporting

During client testing, it was found that there were missing columns required when generating a csv for financial reporting purposes. The sbc_staff_invigilated, start_time, end_time, number_of_students and notes fields were added to the generated csv file.

This PR also attempts to reduce the complexity of the exam_export_list.py to meet sonarqube specifications by reducing complex/nested if/else statements and increasing modularity by creating smaller working functions.

* Need Collect Fees Yes/No

Client testing found that the Collect Fees dropdown on the edit booking modal only had the "No" value enabled, and required a "Yes" value as well. This field had to map correctly to the fees field on the booking model. This PR adds this dropdown option and also ensures that the Put method to edit a particular booking works for this field now that it has been enabled.

Client also asked that the the 3 fields (fee amount, invoice to, name of entity to bill) that were enabled when the fees option was set to yes, were also to be removed, to be possibly implemented at a later date.

* Finalized Postman Tests

Bookings application required finalized postman tests before passing off to the client.

* Filter out Exam Returned - 90 Days or Older

Client testing found that the Exam Inventory table needed to filter out exams that were older than 90 days. This was implemented by altering the exams_list end-point to union two queries: one that fould find all exams where the exam_returned_date was null, and another where the exam_returned_date was greater than 90 days, then unioning the two subqueries together.