Added Ability to Schedule "Other" exams to offsite locations
Added onsite/offsite question to the add Other Exam form workflow with default value set to onsite.  Selecting offsite causes the exam to be posted with offsite_location = '_offsite' which then causes the exam to be booked in the Edit Group Booking Modal when choosing the Scheduling Action later from the exam inventory table.

Adjusted the logic in the exam inventory table to facilitate this new option.