Sprint 10 2 (#222)
* Pesticide Exam Modal - Preset/Validate Exam for Pesticide Exam Type

Client testing found that the Pesticide Exam modal needed the exam type value to be preset to "Pesticide". This PR contains the fix for this issue by pre-setting the exam_type_id once the modal is opened, and validating the exam_type_id as well, such that the user can continue through the steps for creating an exam.

* Admin Panel - Designate Label Re-name

Client testing found that ITA, Pesticide, and Liaison designates needed to be re-named to properly reflect the csr's job title. This has been addressed in this PR.

- ITA -> Office Exam Manager
- Pesticide -> Pesticide Client Liaison/Program Specialist
- Liaison -> ITA Liaison/Program Specialist

* Responsive Exam Inventory Table

Client testing found that if a user with vision impairment zooms in on the exam inventory table, that it didn't have a horizontal scroll automatically enabled. This was fixed by adding the 'responsive' prop to the b-table on the exam-inventory table component, and tested by zooming on the component to see whether or not a scroll bar became active if the component was larger than the actual screensize of the user.

* Booking Calendar - Booking Events in the Past

Client testing found that bookings could be done in the past, which should not be allowed. This PR corrects this by adding logic to the config object of the calendar, checking to see if the start date of the event being created is before a datetime object called "Today" which goes down the minute value of the datetime object created. Example: If the time of the object created is 3:03pm of that day, the calendar will not allow you to create an event until the 3:30pm-4:00pm block.

* Group Exam Time Select

Client testing found the "Ok" confirm at the bottom of the Exam Time select drop down was not necessary. This PR seeing the confirm option on the Time Question component removed to remedy this.

* Input Field - Disable Autocomplete

Client testing found that vue input fields should have autocomplete disabled.

This was completed in a previous sprint, but developer testing found one field on the edit booking modal that didn't have this setting enabled.

* Generate CSV - GA Reporting

During client testing, it was found that there were missing columns required when generating a csv for financial reporting purposes. The sbc_staff_invigilated, start_time, end_time, number_of_students and notes fields were added to the generated csv file.

This PR also attempts to reduce the complexity of the exam_export_list.py to meet sonarqube specifications by reducing complex/nested if/else statements and increasing modularity by creating smaller working functions.

* Need Collect Fees Yes/No

Client testing found that the Collect Fees dropdown on the edit booking modal only had the "No" value enabled, and required a "Yes" value as well. This field had to map correctly to the fees field on the booking model. This PR adds this dropdown option and also ensures that the Put method to edit a particular booking works for this field now that it has been enabled.

Client also asked that the the 3 fields (fee amount, invoice to, name of entity to bill) that were enabled when the fees option was set to yes, were also to be removed, to be possibly implemented at a later date.

* Finalized Postman Tests

Bookings application required finalized postman tests before passing off to the client.

* Filter out Exam Returned - 90 Days or Older

Client testing found that the Exam Inventory table needed to filter out exams that were older than 90 days. This was implemented by altering the exams_list end-point to union two queries: one that fould find all exams where the exam_returned_date was null, and another where the exam_returned_date was greater than 90 days, then unioning the two subqueries together.

* Object Not Iterable (#216)

Client testing found a bug in the exam_list endpoint when a csr had the liaison designate ind set to 1. The .all method was removed from the query object to resolve this.`

* NPM Babel Versioning Broken Jest

Throughout development of the bookings application, it was found that different versions of babel-core(@babel/core) were being using, which caused a conflict not allowing jest tests to run properly. This was fixed by removing babel-core from the package.json file, updating the file with @babel version 7 packages, and re-installing the packages using npm install. After this was completed jest tests from theq were found to work once the proper environment variables were set.

* SQL Script Clean-up

Client found issues with the manage script after rebuilding the database on the test server.

* Q-Service Appointments Booking

Added another full-calendar based component to the application for handling appointments for services provided at serviceBC offices through the existing Q application.  Citizens are booked by name into 15 minute slots on the branch Appointent Calendar and when they arive, they can be checked directly into the Q.

Also fixed exam inventory table min height issue and defaulted the booking calendar to show offsite.

* Postman Environment Variables

Client pointed out that refreshed enviornment variables weren't sent on the lastest PR. These were changed over a month ago with the move to localhost keycloak, but forgotten.

* Fixed Issue with Add Pesticide Workflow

Fixed the issue causing the office_id to not validate during pesticide exam capture preventing the submission of any pesticide exams.

* Added Ability to Schedule "Other" exams to offsite locations

Added onsite/offsite question to the add Other Exam form workflow with default value set to onsite.  Selecting offsite causes the exam to be posted with offsite_location = '_offsite' which then causes the exam to be booked in the Edit Group Booking Modal when choosing the Scheduling Action later from the exam inventory table.

Adjusted the logic in the exam inventory table to facilitate this new option.