Exam Type Filtering - Group Pesticide Edge Case
During client testing, a group pesticide exam was created, which showed up in both the group exam and pesticide exam add exam modals. To have this exam only show up in the Group Exam modal, the filtering for exam types of the pesticide modal was altered to exclude any exams with the word 'Group' in them.