Sprint 10 3 (#225)
* Ensure that edit exam modal closes when we delete an exam

* Generate Financial Report Button Access

During the handoff meeting, clients determined that this button should now be viewable by users with the ita_designate and financial_designate fields set to 1 and if a user is also a GA. This button was also modified to no longer be justified right on the button row, and is now in line with the rest of the buttons that show up on the exam inventory page.

* Exam Type Admin Panel - Ind Re-name

Client testing found that columns labels with "_ind" were confusing (ind->individual, not ind->indicator), so this was changed to label the list table and table column fields to show 'Flag' instead of 'Ind'.

* Chanllenger Exams - Show if User is Office Manager

During final client testing, it was found that the challenger/monthly sessional drop down needed to be visible to users with the office manager flag set to 1. This was resolved by adding the 'is_ita_designate' getting to the logic that controls whether or not this dropdown component is visible.

* Exam Type Filtering - Group Pesticide Edge Case

During client testing, a group pesticide exam was created, which showed up in both the group exam and pesticide exam add exam modals. To have this exam only show up in the Group Exam modal, the filtering for exam types of the pesticide modal was altered to exclude any exams with the word 'Group' in them.

* Dynamic Postman Environment Variables

Postman tests for the bookings application now supports dynamic environment variables, such that the user that runs the test no longer has to re-build their backend each time they run the test suite.

* Calendar Event Width

Clients asked for the calendar events to show up with ~90% width of the calendar room column to allow users to place more than one event in the same time slot. I've confirmed that the 85% width the event is currently set to is satisfactory with the client already.

* Exam Types - Exam Duration with Minutes

Client testing found that pesticide exams had durations that weren't round number of hours (ie// 1h30m, 1h45m). To accomodate for this, a number_of_minutes field was added to the exam_type model/schema and then added to booking durations in the calendar, booking-modal and edit-booking-modal vue files. This was tested with the client over zoom to ensure that there was no confusion on how exams were created and handled.

* Revert "Pesticide Exam Modal - Preset/Validate Exam for Pesticide Exam Type"

This reverts commit 12d401421de2c7651dc5791f4567aa5feb7293f6.

* Pesticide Exam - Revert Exam Type Changes

Client has asked that the default exam type change that was made in commit
12d4014 be reverted since there are now going to be 3 different pesticide exam types (accounting for duration). A sort of the exam types has also been applied such that the list will now mirror functionality of all other exam type selections in all other add exam modal types.

* Fixed the issue causing the @ not to load in IE at all

Elimanated the legacy Babel 6-style configuration files for Babel (frontend/.babelrc) in favour of a Babel 7 style babel.config.js file.  Appropriately specified targets and environments necessary to support IE11 and Firefox/Chrome in all the required NODE_ENVs

* Fixed Alembic version issue

* Fixed migrations to work with upstream changes