Sprint 10 5 (#232)
* Jest Tests - Booking an Exam

This PR contains the jest test responsible for booking an exam. It first creates an individual exam, then books the exam on the day after the current date. This test is meant to be run with an empty booking table, and can be run with exams in the exam inventory table. To run this test type `npm test -t bookings-book-exam` in command line.

* Two Add ITA Exam Buttons

Client testing found that if a user had the Office Manager permission set to 1 and wasn't a GA, that two Add ITA Exam buttons would show on the exam invetory page. This issue was fixed by fixing partial logic on the buttons-exams such that a negation is now complete.

* Notification Z-Index

Client testing found that the notification banner was being rendering over top of modals. This was fixing by reducing the the z-index of the notification banner from 1041 to 1040.

* Need Collect Fees Yes/No

Client testing found that the Collect Fees dropdown on the edit booking modal only had the "No" value enabled, and required a "Yes" value as well. This field had to map correctly to the fees field on the booking model. This PR adds this dropdown option and also ensures that the Put method to edit a particular booking works for this field now that it has been enabled.

Client also asked that the the 3 fields (fee amount, invoice to, name of entity to bill) that were enabled when the fees option was set to yes, were also to be removed, to be possibly implemented at a later date.

* Click Once Submit Buttons

Client noted that during slow network periods that clicking submit buttons on create exams or create bookings would create multiple instances of that object. This has been resolved by adding '.once' to the @click property of the submit buttons on the add exam modal and add booking modal.

* Typo

Found a typo in the exam success notification wrt the setEditExamSuccess mutation

* Fixed Issue that prevented booking of session exams in office

As the AddExamModal received no props, eliminated the need to save its state into another store key before switching routes to use the Calendar.  Now that the AddExamModal persists and is not destroyed on route changes there is no need to save and re-load its state.  Removed numerous actions, mutations, state keys and methods accross AddExamModal, BookingModal, AddExamModule and AddModalFormComponents.  Booking on-site now works.