Modified EditGroupBookingModal to Support Booking Other Exams Offsite
Changed guards to allow CSRs as well as GA/Designates to edit all fields for the "other" exam type and removed the special case fields being displayed in favour of just showing the date and time pickers but with conditions on disabled prop.

Also fixed in issue where Add Exam could fail due to submitting without an exam_received_date.