Fix IE Issue: EditExamModal datepicker does not update on change
Component local state storing values in object as keys which were initialized if they had preset values but exam_received_date is not set in this case and so the v-model was creating the key non-reactively.  Added exam_received_date: null to the predifined component local data to ensure reactivity.