Admin Panels - Delete Flag Based on User Roles
Due to use in production, clients found that the can_delete flag needed to be added to the room and invigilator panels due to flask admin ignoring FK contraints when deleting objects. The rooms panel now no longer has the ability to delete whatsoever, and the invigilator panel now selectively shows the delete button based on the users role.