Sprint 11 2 (#259)
* Updated finance reporting now that individual exams can be booked offsite

* Financial Reporting - Filtering Offices

Clients asked that if a user is a financial reporting designate to return ALL exams when the query for reporting is submited. Otherwise the exams returned to the user are filtered based upon the users office.

* Fix Edit Exam Modal: Date Received / Expiry Date Inconsistency

App was inconsistent with formatting/parsing of dates at capture, display and during POST/PUT actions.  All dates now captured in local time, submitted as UTC and converted to local time for display.  This eliminates the issue where DatePickers were showing the day before the actual date value held in the backend/store

* Fixed Issue w/ the Q where CSR able to navigate away when serving citizen

Added condition to display of hamburger menu to prevent this from happening

* Fix Serve Citizen Modal Loading Spinner

Previously added trigger to show the spinner when clicking on a citizen from the queue table but failed to consider the scenarios of clicking on a citizen in the hold table or directly from the Add Citizen Modal.  Added triggers for the other ways to launch the modal.

* Group Exam Notifications - SBC Staff

Clients noticed after production release that group exams were causing the notification banner to show if SBC Staff were selected as invigilators. This condition met the requirements of not having to notify staff about exams, and such was fixed by adding a filter the group exam object in the csrs/me endpoint that only counted exams with booking.sbc_staff_invigilated == 0. This was tested by creating a group exam object, then creating a booking for this exams without an inviglator and making sure the notification would present itself on refresh. Then, altering the booking to have SBC staff invigilate the booking, refreshing the page to make sure no notification was present. Finally, the booking was altered to have an actual invigilator present for the booking, and refreshing the page to make sure that the notification banner was not present.