Sprint 11 4 (#261)
* Fixed Edit Group Exam Modal to Disable Past Date Selection

Exam dates can no longer be edited to a date that is in the past.  Also fixed eratic form validation that prevented the submit button from being pressed when certain dates were selected.

* Fixed Issue with Appointment Booking Modal Notes

Notes were being captured but not displayed in the modal.  Fixed.  Also adjusted calendar start time to 8:30am.

* Fixed Errors in Branch Agenda

Previously this component was more complex and it relied on an invisible full-calendar instance to consume and parse the event data however this approach was causing issues.  Since the Agenda has recently been simplified and no longer supports multiple views etc., there is no need for full-calendar.  Have re-implemented using only moment.js and eliminated the errors related to full-calendar

* Deleted Date for Invigilators and Rooms

Client was flooded with requests for how to delete invigilators and rooms upon launch. To accomodate for this request but maintain data integrity, a deleted date was added to both the invigilator and room models, such that if these fields were set to a date in the admin panels for these objects, they will no longer show up in the application drop downs or calendars.

* Allow Office Mgr to Edit Group/Session Exam Details

Logic controlling access to the full version of the Edit Exam Modal did not consider the ita_designate key.  Office Manager previously saw only the fields-limited CSR version.   Changed so that office mgr has same access as GA role.  Also fixed:
-liaison_designate were unable to edit office of pesticide exams.
-the exam_type field of the EditExamModal was listing exam_types which did not make sense. Disabled the ability to switch exam_type on a pesticide exam or challenger exam at all and restricted changing 'other' 'individual ita' and 'group ita' to exam type options of their catagory only.
-adjusted the other field-display and field-disable logic to ensure access to necessary features for office managers
-set the time of the editExamSuccess banner to 3 seconds instead of 6

* Admin Panel - Room -> Office Name Sort

Clients required rooms to be sorted by office name after testing in production.