Delete Room Admin Warning
Clients required there to be a warning given to admin users if a room is about to have a deleted date applied to it, but it currently being used for bookings in the future. The flask admin on_model_change method was used to detect changes in the deleted date field before submission, and looking at the booking model for instances of such a room being used with start times greater than today (doesn't care about historical instances). If the room isn't in use, the room can be deleted. If the room is in use, and error message propagates telling the user that they were unable to apply the deleted date. Any other changes that the user might make on top of applying a deleted date to a room in use will be saved.