Sprint 11 6 (#263)
* Admin Panel - Room Searching

Client required rooms to be searchable through admin panels.

* Backend Changes for Tracking Number Length

Bookings application required backend changes to reflect tracking number length of 255 chars. Alembic would not generate a migration file for the length change, so the last migration file in the alembic chain was modified to see this change happen.

* Return Exam fix: Action Taken Length & Deleted Invigilator Changes
-Action Taken field now displays an error at the maximum input length and prevents further input
-Adjusted the exam_inventory table to display the life-ring error symbol on any exam assigned to a deleted invigilator and adjusted the filters to show such exams as not ready.
-Changed a call to previous method name which was revised in EditGrouoBooking modal

* AddExamForm Modal exam_name length validation

Modified existing validation logic to check if exam_name field exceeds 50 characters in length and displays error / prevents advancing until corrected.

* Added exam_name length validation to EditExamModal

Added validation and an error message as per the Return Exam modal at the 50 character limit.

* Fixed Error in Appointments with double clicking

Fixed error where double clicking the appointments page would cause two modals to open.

Also removed stray console.log from return_exam_modal

* Delete Room Admin Warning

Clients required there to be a warning given to admin users if a room is about to have a deleted date applied to it, but it currently being used for bookings in the future. The flask admin on_model_change method was used to detect changes in the deleted date field before submission, and looking at the booking model for instances of such a room being used with start times greater than today (doesn't care about historical instances). If the room isn't in use, the room can be deleted. If the room is in use, and error message propagates telling the user that they were unable to apply the deleted date. Any other changes that the user might make on top of applying a deleted date to a room in use will be saved.