Sprint 11 2 (#265)
* Sprint 11 4 (#261)

* Fixed Edit Group Exam Modal to Disable Past Date Selection

Exam dates can no longer be edited to a date that is in the past.  Also fixed eratic form validation that prevented the submit button from being pressed when certain dates were selected.

* Fixed Issue with Appointment Booking Modal Notes

Notes were being captured but not displayed in the modal.  Fixed.  Also adjusted calendar start time to 8:30am.

* Fixed Errors in Branch Agenda

Previously this component was more complex and it relied on an invisible full-calendar instance to consume and parse the event data however this approach was causing issues.  Since the Agenda has recently been simplified and no longer supports multiple views etc., there is no need for full-calendar.  Have re-implemented using only moment.js and eliminated the errors related to full-calendar

* Deleted Date for Invigilators and Rooms

Client was flooded with requests for how to delete invigilators and rooms upon launch. To accomodate for this request but maintain data integrity, a deleted date was added to both the invigilator and room models, such that if these fields were set to a date in the admin panels for these objects, they will no longer show up in the application drop downs or calendars.

* Allow Office Mgr to Edit Group/Session Exam Details

Logic controlling access to the full version of the Edit Exam Modal did not consider the ita_designate key.  Office Manager previously saw only the fields-limited CSR version.   Changed so that office mgr has same access as GA role.  Also fixed:
-liaison_designate were unable to edit office of pesticide exams.
-the exam_type field of the EditExamModal was listing exam_types which did not make sense. Disabled the ability to switch exam_type on a pesticide exam or challenger exam at all and restricted changing 'other' 'individual ita' and 'group ita' to exam type options of their catagory only.
-adjusted the other field-display and field-disable logic to ensure access to necessary features for office managers
-set the time of the editExamSuccess banner to 3 seconds instead of 6

* Admin Panel - Room -> Office Name Sort

Clients required rooms to be sorted by office name after testing in production.

* Typo (#262)

Missing command from sprint-11-4 PR.

* Sprint 11 6 (#263)

* Admin Panel - Room Searching

Client required rooms to be searchable through admin panels.

* Backend Changes for Tracking Number Length

Bookings application required backend changes to reflect tracking number length of 255 chars. Alembic would not generate a migration file for the length change, so the last migration file in the alembic chain was modified to see this change happen.

* Return Exam fix: Action Taken Length & Deleted Invigilator Changes
-Action Taken field now displays an error at the maximum input length and prevents further input
-Adjusted the exam_inventory table to display the life-ring error symbol on any exam assigned to a deleted invigilator and adjusted the filters to show such exams as not ready.
-Changed a call to previous method name which was revised in EditGrouoBooking modal

* AddExamForm Modal exam_name length validation

Modified existing validation logic to check if exam_name field exceeds 50 characters in length and displays error / prevents advancing until corrected.

* Added exam_name length validation to EditExamModal

Added validation and an error message as per the Return Exam modal at the 50 character limit.

* Fixed Error in Appointments with double clicking

Fixed error where double clicking the appointments page would cause two modals to open.

Also removed stray console.log from return_exam_modal

* Delete Room Admin Warning

Clients required there to be a warning given to admin users if a room is about to have a deleted date applied to it, but it currently being used for bookings in the future. The flask admin on_model_change method was used to detect changes in the deleted date field before submission, and looking at the booking model for instances of such a room being used with start times greater than today (doesn't care about historical instances). If the room isn't in use, the room can be deleted. If the room is in use, and error message propagates telling the user that they were unable to apply the deleted date. Any other changes that the user might make on top of applying a deleted date to a room in use will be saved.

* Changed displayed hours of Appointments Calendar to 8.30 am - 5 pm (#264)