Generate Financial Report - Non Exam Events
Client required reporting to be done on Non Exam Events in the booking model as well as Exam events. A separate query was added to the exam_export end point for booking events without exam IDs, and then added another for loop to the query return results parsing, to parse this object and insert rows in the same csv in the form of exam events. Confirmed with the client beforehand what fields should be blank, which fields required hard coded values and what columns were net new (fees collected)