Service Ctzn Mnmize Button now hides Modal + Allows Navigation
Changed the Minimize Button of the Serve Citizen Modal to hide the modal (on the condition that the Service Begun button has been pressed.  If citizen is 'invited' only, will revert to previous collapse behaviour).

Added conditions to Dash Table, Hold Table, router Navigation Guard for automatic ticket creation in exams, and Dash Buttons, to prevent starting additional tickets with current ticket minimized.

Adjusted Simplified Add Service Modal to display the catagories option but default it to 'Exams' instead of hiding it and fixing it on Exams only.  Adjusted the Serve Citizen modal to allow creation of ticket from Outside the Q and display message indicating noa active ticket when there is none.