Generate Financial Reporting - Bug Fixes
Client reported that financial reporting flag wasn't working properly. This has been fixed and tested to ensure that it's working as intended.

Client asked that there be an option added for ALL BOOKING EVENTS to the generate modal, and that there be an ALL EXAMS option as well. Not merging both options. This has been remedied and the UI to the modal has been updated to have a dropdown with filter options in the dropdown.

Client reported that the ITA and Non-ITA filters were reporting appointments and not just exams. This has been fixed and tested.

During testing, an issue was found with creating rooms in the admin panel. This was fixed being utilizing the `is_created` param in the on_model_change function to check whether or not changes are happening during an EDIT or a CREATE.