Modified TheQ to show appointment details
Appointments module now saves some appointment details to the citizen_comments key.  Modified the DashTable and DashHoldTable to display this data if present.  Modified the ServeCitizen modal to parse out the data on retrieval so it does not show in the comments box and parse it back in on submission so it is retained.