Sprint 11 3 (#273)
* Code Smells - Services End Point

Sonar pointed out that the services end point had an unused variable in it. After further inspection, there were also 2 unused import statements, improper white spacing and misuse of multi line commenting. All of these issues were addressed in this PR.

* Code Smells - CSR Schema

Minor code smell found on the CSR schema where commented code wasn't being used. This was removed because it's useless to have.

* Code Smells - Agenda

Sonar found that mapgetters wasn't being used in the agenda.vue file. This was removed in this PR.

* Code Smells - Appointments

Sonar pointed out that there were 2 unused imports in the appointments.vue file. This was addressed in this PR.

* Code Smells - Appointments

Sonar pointed out that there was one unused import statement, as well as one unused variable in the appt-booking-modal.vue file. This was addressed in this PR.

* Code Smells - Add Exam Form Components

Sonar pointed out that mapActions was not being used in the add-exam-form-components.js file. This was removed in this PR.

* Code Smells - Exam Buttons

Sonar brought to our attention that mapActions was being called, but never used in the buttons-exams.vue file. This was removed in this PR.

* Code Smells - Office Drop

Sonar pointed out that there was a set of 7 unused lines in the office-drop.vue file. These lines correspond to an unused local variable 'officeSet' which is never called or used anywhere else in this file.

* Code Smells - Appointments Module

Sonar pointed out that an unused import of vue was being used in the appointments-module.js file. This was removed in the PR.

* Fix For Issue with Booking Calendar Title after using Agenda

Listners for agenda events were somehow not reliably being destroyed and methods were still  being invoked by identical emitted messages from other calendar component.  changed the emit messages so there is no clash.

* Fix for disapearing hamburger menu on exams after refresh

Changed the logic for displaying of the hamburger to fix this issue.