Multiple Appointment Bookings
This PR addresses client feedback where only one appointment was allowed to be scheduled per given timeslot available in an office. The appointments calendar effectively works like the bookings calendar now, with the calendar background behind a scheduled event being made clickable now. Initial width of a given event in the column is 85% of it's max-width, but after a second event is scheduled, the remaining clickable column width returns to 5%.