Phase 1 - Group/Pesticide Exam Office Validation
Client noticed that when Group exams were created in series, the Office Name/Number fields were pre populated with the previous created exams value, instead of being blank. Also, validation of the field was already happening if the field was blank, which was definitely a bug. This bug was resolved by removing the office_id value being set on initialization of the add exam modal for group and pesticide exams.