Phase 1 - Reset Invigilator When Booking is Rescheduled.
Clients required that when a booking is rescheduled, that the invigilator for that booking is removed. In order to do this, a rescheduling flag was added to the edit-booking-modal, such that if the rescheduling button is pressed, this flag is set to true. When the modal is shown, logic was added such that if the flag is true, the invigilator object is set to null which changes the invigilator drop down to null, and when the modal is submitted the invigilator id for that booking is set to null.