Phase 1 (#287)
* Multiple Appointment Bookings

This PR addresses client feedback where only one appointment was allowed to be scheduled per given timeslot available in an office. The appointments calendar effectively works like the bookings calendar now, with the calendar background behind a scheduled event being made clickable now. Initial width of a given event in the column is 85% of it's max-width, but after a second event is scheduled, the remaining clickable column width returns to 5%.

* Phase 1 - Active Citizen Count

Clients required that the count of citizens waiting in queue be present on the page for users based on the following:
	- if the office was a reception office (smart board type: nocallonsmartboard)
	- if the route of the app is either 'exams' or 'bookings'

Based on this, logic from the Q route was re-used to display the citizen waiting count now in the application header.

* Phase 1 - Non Exam Booking Contact Information

Clients noted that when booking multiple back-to-back non-exam events in the bookings calendar, that the contact information from the previous booking would carry forward.

This bug was resolved by setting the contact information variable for the other booking to modal to be an empty string after the POST has completed.

* Phase 1 - Pesticide Exam Candidate Name

Clients noticed that ITA Liaison users were unable to edit candidate names for pesticide exams. The logic that controlled this field on the edit exam for did not include the pesticide exam type, and was edited to include this exam type.

* Phase 1 - Group/Pesticide Exam Office Validation

Client noticed that when Group exams were created in series, the Office Name/Number fields were pre populated with the previous created exams value, instead of being blank. Also, validation of the field was already happening if the field was blank, which was definitely a bug. This bug was resolved by removing the office_id value being set on initialization of the add exam modal for group and pesticide exams.

* Phase 1 - Reset Invigilator When Booking is Rescheduled.

Clients required that when a booking is rescheduled, that the invigilator for that booking is removed. In order to do this, a rescheduling flag was added to the edit-booking-modal, such that if the rescheduling button is pressed, this flag is set to true. When the modal is shown, logic was added such that if the flag is true, the invigilator object is set to null which changes the invigilator drop down to null, and when the modal is submitted the invigilator id for that booking is set to null.