Phase 1 - Notification Banner and Exam Inventory Filter Updates
Client required that the notification banner for group/individual exams had the ability to re-direct the user to the exam-inventory page WITH preset inventory filters set. This PR has a button added to the rightmost corner of the notification banner, that when clicked, pushes the user to the exam route, and sets the exam filter mutation to specific values based upon which exam notification is present (group, individual or both).

The filters on the Exam Inventory have been consolidated to two dropdowns: one that contains exam types (ind, group all) and another that has present filter combinations. Extended filtering functionality was added to the filter function present on the table to account for BOTH individual and group exams that require attention, as there was no way to accurately display the combined set of both exam types that require attention.

All present functionality on the exam inventory page was tested after filtering was changed to ensure that new changes only affected the scope of this pull request.