Phase 2 - Database Enhancements
Since moving the queue application from a mysql backend to a postgres backend, the client has noticied far too many long running queries and timeouts on the backend. This prompted a review/audit of how and what the database is returning when API calls are made.

The first thing that was reviewed was the schema objects for exams and bookings. Here, only very specific rows were returned, cutting down the size of the JSON object returned in most cases by half, which in turn sped up query return times greatly. Next the lazy parameters on the exam <-> exam_type relationship was reviewed. First the lazy condition was removed, then altered to be 'raise'. This removed an outstanding long query response time from logs. Finally, a deleted_date filter was added to the exam_put end-point which helped reduce return time for this action as well.

All actions (Create, Read, Update and Delete) were reviewed during and after work was completed for any application functionality that touched the exam or booking end-points that use the models/schemas/end-points that were changed.