Phase 1 - Office Name Field Reset
Client found a small bug with previous phase 1 work, where if the user enters an exam via add exam modal, and clicks the "Log Another" button in the final step, the office name and number field is re-populated with the previous office name and number field.

Additional functionality had to be added to the logAnother function to reset the state object when the form is re-rendered to have a blank office_number field.