Phase 2 - Booking and Appointment Blackouts
This new feature allows users with role code of GA to schedule blackout events in both the booking and appointment calendars. These events are denoted in their respective calendars with a BLACK color. Currently the Appointment blackout modal maps 1 event to 1 day. The booking blackout modal has the ability to map 1 event to 1 room in 1 day as well as 1 event to many rooms in 1 day.

Future Work:
	- Ability for Support Role to blast events to many or all offices
	- Ability to push booking <-> appointment blackout information seamlessly
	- Ability to create recurring blackout events