Group Exam Event ID Warning
Client required that the event ID field on GROUP EXAM CREATE display a warning if the event ID entered into the form was already in use.
This was accomplished by creating a new endpoint to look at exam event ids and return a true or false value based upon whether or not
the current value of that event_id is present in the database. The validate function already in place fires on every keystroke, so if
the event id is X chars long, the endpoint is hit X times in order to validate.