Phase 4 items new (#337)
* Group Exam Event ID Warning

Client required that the event ID field on GROUP EXAM CREATE display a warning if the event ID entered into the form was already in use.
This was accomplished by creating a new endpoint to look at exam event ids and return a true or false value based upon whether or not
the current value of that event_id is present in the database. The validate function already in place fires on every keystroke, so if
the event id is X chars long, the endpoint is hit X times in order to validate.

* Recurring Events Edit/Delete Series - Booking

Client required functionality to edit and delete recurring events on the booking calendar(non-exam and blackouts). This required adding new end points to the booking package for editing and deleting recurring events.

Front end development shows a new row on the edit booking modal, only visible if the booking event has a recurring_uuid associated with it. This row houses a button to delete the recurring series, and a toggle button for editing
the series of events or a single event.

* Recurring Events Edit/Delete Series - Appointments

Client requires functionality in order to edit and delete a series of recurring events. This commit edits functionality behind the appointments calendar/modals in order to do this.

Two new end points were created in order to accomplish. Both require a uuid to be provided, which is now generated on creation and submitted on creation to the updated appointment model.

* Non-Recurring Exams - Pre-populate Start time and date fields with calendar click

Client required functionality for the non-exam recurring events form to be pre-populated with the start time and date of when/where the calendar was clicked.

This pre-population broke the generateRule functioni, which had to be rebuilt to handle timezones more robustly than other recurring modals.