Phase 4 - Booking Delete History
For reporting purposes, booking recurring events will not be deleted after the current date of when the delete function is being called.

This PR also contains a fix for a bug found where if the edit series toggle is turned on, the rescheduling button is hidden along with the date and time incrementer.